Some code changes
Cleaned up codes, removed redundant functions and variables, removed security features, removed all print and write statements, now the controller is dumb and will only do what the computer tells it to do, and it will not respond back once it did what the computer told it to do. This is great for simpler stuff that don't complex features.